backend/requirements.txt
Flask==2.2.5
python-multipart==0.0.5
pdfplumber==0.11.7
python-docx==0.8.11
spacy==3.5.4
sentence-transformers==2.2.2
torch==2.2.0
werkzeug==2.2.3
gunicorn==21.2.0